+++ idea/src/org/jetbrains/kotlin/idea/refactoring/introduce/extractClass/ui/KotlinExtractInterfaceDialog.kt
@@ -62,7 +62,9 @@ class KotlinExtractInterfaceDialog(
                 val member = memberInfo.member
                 return !(member.hasModifier(KtTokens.INLINE_KEYWORD) ||
                          member.hasModifier(KtTokens.EXTERNAL_KEYWORD) ||
-                         member.hasModifier(KtTokens.LATEINIT_KEYWORD))
+                         member.hasModifier(KtTokens.LATEINIT_KEYWORD) ||
+                         member.hasModifier(KtTokens.INTERNAL_KEYWORD) ||
+                         member.hasModifier(KtTokens.PROTECTED_KEYWORD))
             }
 
             override fun isAbstractEnabled(memberInfo: KotlinMemberInfo): Boolean {